vma: use BlockBackend on extract
As we else rely on bdrv_close_all() do clean up, which was rewritten
in ca9bd24cf1d53775169ba9adc17e265554d1afed and fails on "dangling"
BDS pointers, such a pointer exists with *bs.
Use the BlockBackend to get our BDS and just unref the BlockBackend
when done, it handles the rest for us.

The other two calls to bdrv_close_all() happen in verify_content()
and dump_config(), both do not have a BDS so no need to change here.