rbd: fix cache mode behavior
Either the cache mode asks for a cache or not. There's no
point in having a "temporary" cache mode. This option AFAIK
was introduced as a hack for ancient virtio drivers. If
anything, we should have a separate option for it. Better
yet, VMs affected by the related issue should simply
explicitly choose writethrough.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>